Split up and reorganized with a new api.
